Change memory parameter type from uint64_t to uint32_t

+++ src/windows/WslcSDK/wslcsdk.cpp
@@ -34,7 +34,7 @@ STDAPI WslcSessionSettingsSetCpuCount(_In_ WslcSessionSettings* sessionSettings,
     UNREFERENCED_PARAMETER(sessionSettings);
     return E_NOTIMPL;
 }
-STDAPI WslcSessionSettingsSetMemory(_In_ WslcSessionSettings* sessionSettings, _In_ uint64_t memoryMb)
+STDAPI WslcSessionSettingsSetMemory(_In_ WslcSessionSettings* sessionSettings, _In_ uint32_t memoryMb)
 {
     UNREFERENCED_PARAMETER(memoryMb);
     UNREFERENCED_PARAMETER(sessionSettings);
